requirements.txt
selenium==3.141.0
behave===1.2.6
allure-behave==2.8.16
PyHamcrest==2.0.2
click==7.1.2
prettytable==2.0.0
psutil==5.7.3
autopep8==1.5.3
pylint==2.5.3
